+++ src/tests/mod.rs
@@ -70,4 +70,5 @@ mod test_unrelated_module_cache;
 mod test_util;
 mod test_valgrind_suppression;
 mod test_value_splitting;
+mod test_version_report;
 mod test_wide_return_tail_call;